api-gateway/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
PyJWT==2.12.1
httpx==0.25.0
prometheus-fastapi-instrumentator==7.1.0
python-multipart==0.0.22
anyio==4.13.0
starlette==0.52.1
